+++ src/units/coordinate.rs
@@ -1,4 +1,4 @@
 pub struct Coordinate {
-    pub latitude: f64,
-    pub longitude: f64,
+    pub latitude: f32,
+    pub longitude: f32,
 }
